Adds integration test for custom measurement unit and default breakpoints
