+++ cypress/integration/custom-configuration/custom-breakpoints.test.js
@@ -1,0 +1,16 @@
+it('Supports custom breakpoints and measurement unit', () => {
+  cy.visit('/custom-configuration/custom-breakpoints')
+
+  const assertPadding = (values) => {
+    return cy.get('#composition').should('have.css', 'padding', values)
+  }
+
+  assertPadding('16px')
+  cy.setBreakpoint({
+    minWidth: 576,
+    maxWidth: 768,
+  }).then(() => assertPadding('32px'))
+  cy.setBreakpoint({
+    minWidth: 769,
+  }).then(() => assertPadding('48px'))
+})
