+++ cypress/integration/custom-configuration/custom-unit.test.js
@@ -1,0 +1,6 @@
+import assertBell from '../../asserts/assertBell'
+
+it('Supports custom measurement unit with default breakpoints', () => {
+  cy.visit('/custom-configuration/custom-unit')
+  assertBell()
+})
