+++ cypress/integration/custom-configuration/index.js
@@ -1,0 +1,4 @@
+describe('Custom configuration', () => {
+  require('./custom-unit.test')
+  require('./custom-breakpoints.test')
+})
