+++ cypress/integration/index.js
@@ -1,3 +1,4 @@
 require('./behavior')
 require('./responsive-props')
 require('./composition')
+require('./custom-configuration')
