+++ cypress/integration/responsive-props/index.js
@@ -2,5 +2,4 @@ describe('Responsive props', () => {
   require('./mobile-first.test')
   require('./breakpoint-specific.test')
   require('./inclusive-bell.test')
-  require('./custom-configuration.test')
 })
