+++ examples/responsive-props/index.jsx
@@ -3,7 +3,6 @@ import { Switch, Route } from 'react-router-dom'
 import MobileFirst from './MobileFirst'
 import BreakpointSpecific from './BreakpointSpecific'
 import InclusiveBell from './InclusiveBell'
-import CustomConfiguration from './CustomConfiguration'
 
 const ResponsiveProps = ({ match }) => (
   <Switch>
@@ -13,10 +12,6 @@ const ResponsiveProps = ({ match }) => (
       component={BreakpointSpecific}
     />
     <Route path={`${match.path}/inclusive-bell`} component={InclusiveBell} />
-    <Route
-      path={`${match.path}/custom-configuration`}
-      component={CustomConfiguration}
-    />
   </Switch>
 )
